refactor: move from request to auth namespace
